@@ -16,7 +16,7 @@
         }
     ],
     "require": {
-        "php": ">=7.1",
+        "php": ">=7.4",
         "illuminate/contracts": "^8.0",
         "illuminate/console": "^8.0",
         "illuminate/database": "^8.0",
